+++ include/linux/splice.h
@@ -48,7 +48,7 @@ struct splice_desc {
 struct partial_page {
 	unsigned int offset;
 	unsigned int len;
-	unsigned long private;
+	uintptr_t private;
 };
 
 /*
